Adds tooltips for easier configuration of widgets
the widgets are styled to "stand in place" after been hovered once. This is done for Accessibility reasons.
There was a problem with the "required" attribute in the 'SCM Markdown Content
'-widget. If this attribute was attached to the the form could no longer be submitted. Therefore, the "required" attribute was only attached to the markdown-path field, but now the markdown-path is only displayed if a branch is available (per angular condition). Ergo, a repository has been selected.
Pinned angular version to 1.8.2.
Changed the package manager from npm to yarn. Therefore added a yarn lock and removed the package-lock.json.
Added .idea folder to gitignore
Update changelog.
Update readme (including the npm to yarn changes).